requirements.txt
# AutoML Debate System — Python dependencies (install from repo root: pip install -r requirements.txt)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0

langgraph>=0.2.50
langchain-core>=0.3.20
langchain-openai>=0.3.0

numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
xgboost>=2.1.0
joblib>=1.4.0
pyarrow>=15.0.0

chromadb>=0.5.0
pysqlite3-binary>=0.5.0

httpx>=0.27.0
